apps/modal-backend/requirements.txt
modal>=0.68
fastapi>=0.115
httpx>=0.27
openai>=1.55
pydantic>=2.9
uvicorn[standard]>=0.32
python-dotenv>=1.0
